+++ src/features/admin/attendees.ts
@@ -36,7 +36,7 @@ import {
   isPaidListing,
   type ListingWithCount,
 } from "#shared/types.ts";
-import { logAndNotifyRegistration } from "#shared/webhook.ts";
+import { logAndNotifyRegistration } from "#shared/webhook/delivery.ts";
 import {
   adminAttendeeDeletePage,
   adminResendNotificationPage,
