+++ src/locales/en/activity-log.json
@@ -5,5 +5,6 @@
   "admin.log.no_activity": "No activity recorded yet",
   "admin.log.recent_entries": "Showing the most recent 200 entries.",
   "admin.log.guide_link": "Activity log guide",
-  "admin.log.square_signature_hint": "Re-enter your Square settings, paying close attention to the name of each field."
+  "admin.log.square_signature_hint": "Re-enter your Square settings, paying close attention to the name of each field.",
+  "admin.log.registration_delivery_failed": "Registration notification delivery failed."
 }
